Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>

+++ sdk/cwl/tests/federation/arvboxcwl/start.cwl
@@ -14,6 +14,9 @@ inputs:
   branch:
     type: string
     default: master
+  arvbox_mode:
+    type: string?
+    default: "dev"
 outputs:
   cluster_id:
     type: string
@@ -71,24 +74,28 @@ arguments:
   - shellQuote: false
     valueFrom: |
       set -ex
-      mkdir -p $ARVBOX_DATA
-      if ! test -d $ARVBOX_DATA/arvados ; then
-        cd $ARVBOX_DATA
-        git clone https://git.arvados.org/arvados.git
+      if test $(inputs.arvbox_mode) = dev ; then
+        mkdir -p $ARVBOX_DATA
+        if ! test -d $ARVBOX_DATA/arvados ; then
+          cd $ARVBOX_DATA
+          git clone https://git.arvados.org/arvados.git
+        fi
+        cd $ARVBOX_DATA/arvados
+        gitver=`git rev-parse HEAD`
+        git fetch
+        git checkout -f $(inputs.branch)
+        git pull
+        pulled=`git rev-parse HEAD`
+        git --no-pager log -n1 $pulled
+      else
+        export ARVBOX_BASE=$(runtime.tmpdir)
+        unset ARVBOX_DATA
       fi
-      cd $ARVBOX_DATA/arvados
-      gitver=`git rev-parse HEAD`
-      git fetch
-      git checkout -f $(inputs.branch)
-      git pull
-      pulled=`git rev-parse HEAD`
-      git --no-pager log -n1 $pulled
-
       cd $(runtime.outdir)
       if test "$gitver" = "$pulled" ; then
-        $(inputs.arvbox_bin.path) start dev
+        $(inputs.arvbox_bin.path) start $(inputs.arvbox_mode)
       else
-        $(inputs.arvbox_bin.path) restart dev
+        $(inputs.arvbox_bin.path) restart $(inputs.arvbox_mode)
       fi
       $(inputs.arvbox_bin.path) status > status.txt
       $(inputs.arvbox_bin.path) cat /var/lib/arvados/superuser_token > superuser_token.txt
